+++ bootstrap.php
@@ -11,6 +11,8 @@ require_once 'functions/template.php';
 require_once 'functions/validators.php';
 require_once 'functions/db.php';
 require_once 'functions/requests.php';
+require_once 'functions/file.php';
+
 if (!file_exists('config.php')) {
     exit('Создайте файл config.php на основе config.sample.php и сконфигурируйте его');
 }
